Moved User management to separate view
