+++ src/main/java/tk/ynvaser/quiz/frontend/view/MainView.java
@@ -27,6 +27,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import tk.ynvaser.quiz.frontend.view.admin.AdminView;
 import tk.ynvaser.quiz.frontend.view.games.ActiveGamesView;
 import tk.ynvaser.quiz.frontend.view.login.LoginView;
+import tk.ynvaser.quiz.frontend.view.usermanagement.UserManagementView;
 
 import java.util.Optional;
 
@@ -92,10 +93,9 @@ public class MainView extends AppLayout {
 
     private Component[] createMenuItems() {
         return new Tab[]{
-                //createTab("Quiz Dashboard", QuizDashboardView.class),
-                //createTab("Quiz (name)", QuizmasterViewView.class),
                 createTab("Adminisztráció", AdminView.class),
-                createTab("Aktív Játékok", ActiveGamesView.class)
+                createTab("Aktív Játékok", ActiveGamesView.class),
+                createTab("User Management", UserManagementView.class)
         };
     }
